% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_creator.R
\name{add_creator}
\alias{add_creator}
\title{Add a "creator" function}
\usage{
add_creator(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{name}{A name for the dataset-creating function}

\item{fn}{A function that creates a simulated dataset}
}
\value{
The original simulation object with the new creator function added
}
\description{
Add a "creator" function to your simulation object.
}
\examples{
# There are two ways to use add_creator(). The first is to declare a function
# and add it to simba later:

sim <- new_sim()
create_data <- function (n) { rpois(n, lambda=5) }
sim \%<>\% add_creator(create_data)

# The second is to do both at the same time:

sim <- new_sim()
sim \%<>\% add_creator("create_data", function(n) {
  rpois(n, lambda=5)
})

# With either option, you can test your function as follows:

sim$creators$create_data(10)
}
