% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{create_levels_grid_big}
\alias{create_levels_grid_big}
\title{Function for creating levels_grid_big}
\usage{
create_levels_grid_big(sim_obj)
}
\arguments{
\item{sim_obj}{A simulation object of class `simba`}
}
\description{
Function for creating levels_grid_big
}
