% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Get the value of a simulation object internal variable}
\usage{
get(sim_obj, variable)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{variable}{The name of the internal variable. Options include the
following:\n
- hey \n
- there}
}
\description{
Get the value of a simulation object internal variable
}
\examples{
sim <- new_sim()
sim \%>\% get("num_sim_total")
sim \%<>\% set_levels(alpha=c(2,3,4))
sim \%>\% get("num_sim_total")
}
