% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complex.R
\name{get_complex}
\alias{get_complex}
\title{Access internal simulation variables}
\usage{
get_complex(sim_obj, sim_uid)
}
\arguments{
\item{sim_obj}{A simulation object of class \code{simba}, usually created by
\link{new_sim}}

\item{sim_uid}{The unique identifier of a single simulation replicate. This
corresponds to the \code{sim_uid} column in \code{sim$results}.}
}
\value{
The value of the complex simulation result data corresponding to the
    supplied \code{sim_uid}
}
\description{
Extract complex simulation data from a simulation object
}
\examples{
sim <- new_sim()
!!!!! TO DO
}
