% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{handle_errors}
\alias{handle_errors}
\title{Function for internal error handling}
\usage{
handle_errors(obj, err, possible = NA)
}
\arguments{
\item{obj}{The object to check for errors (can be a list)}

\item{err}{The type of error to check for (character string)}
}
\value{
Throws and error or returns NULL
}
\description{
Function for internal error handling
}
