% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_support.R
\name{js_support}
\alias{js_support}
\title{Display information about currently-supported job schedulers}
\usage{
js_support()
}
\description{
Run this function to display information about job schedulers
    that are currently supported for running \pkg{SimEngine} simulations on a
    cluster computing system (CCS).
}
\examples{
js_support()
}
