% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the simulation}
\usage{
run(sim_obj, script)
}
\arguments{
\item{sim_obj}{A simulation object created by new_sim()}

\item{script}{The name of a simulation script, added using add_script()}
}
\description{
Run the simulation
}
\examples{
!!!!! TO DO
}
