% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simba.R
\docType{package}
\name{simba}
\alias{simba}
\title{simba: An open-source framework for statistical simulations in R}
\description{
simba is an open-source R package for structuring, maintaining, running, and
debugging statistical simulations on both local and cluster-based computing
environments. Emphasis is placed on thorough documentation, and scalability.
}
\section{Documentation}{

The primary documentation for \pkg{simba} is housed at
\url{https://avi-kenny.github.io/simba} - we recommend that users start here
when first learning how to use the package. Additionally, standard R
documentation is provided, which can be accessed in the usual way (e.g.
\code{?new_sim}).
}

