% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize simulation results}
\usage{
summarize(sim, ...)
}
\arguments{
\item{sim}{A simulation object of class \code{sim_obj}, usually created by
\code{\link{new_sim}}}

\item{...}{One or more named lists, separated by commas, specifying desired summaries of the \code{sim}
    simulation object. Each list must have a \code{stat} item, which specifies the summary statistic to
    be calculated. For all \code{stat} options besides \code{"coverage"}, the \code{name} item is optional;
    if it is not provided, a name will be formed from the type of summary and the column on which the summary
    is performed. The \code{na.rm} item indicates whether to exclude \code{NA} values when performing the calculation (with
    default being \code{FALSE}). Additional required items are detailed below for each \code{stat} type.
    \itemize{
    \item{\code{stat = "mean"}: \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the mean.}

    \item{\code{stat = "median"}: \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the median.}

    \item{\code{stat = "var"} (variance): \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the variance.}

    \item{\code{stat = "sd"} (standard deviation): \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the standard deviation.}

    \item{\code{stat = "mad"} (mean absolute deviation): \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the MAD.}

    \item{\code{stat = "iqr"} (interquartile range): \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the IQR.}

    \item{\code{stat = "min"} (minimum): \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the minimum.}

    \item{\code{stat = "max"} (maximum): \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the maximum.}

    \item{\code{stat = "quantile"}: \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the quantile, \code{prob} is a number in [0,1] denoting the desired quantile.}

    \item{\code{stat = "correlation"}: \code{x} and \code{y} give the names of the variables in \code{sim$results}
    on which to calculate the (Pearson's) correlation.}

    \item{\code{stat = "covariance"}: \code{x} and \code{y} give the names of the variables in \code{sim$results}
    on which to calculate the covariance.}

    \item{\code{stat = "bias"}: \code{estimate} gives the name of the variable in \code{sim$results}
    containing the estimator of interest, \code{truth} is the estimand of interest (see \emph{Details}).}

    \item{\code{stat = "bias_pct"} (percent bias): \code{estimate} gives the name of the variable in \code{sim$results}
    containing the estimator of interest, \code{truth} is the estimand of interest (see \emph{Details}).}

    \item{\code{stat = "mse"} (mean squared error): \code{estimate} gives the name of the variable in \code{sim$results}
    containing the estimator of interest, \code{truth} is the estimand of interest (see \emph{Details}).}

    \item{\code{stat = "mae"} (mean absolute error): \code{estimate} gives the name of the variable in \code{sim$results}
    containing the estimator of interest, \code{truth} is the estimand of interest (see \emph{Details}).}

    \item{\code{stat = "coverage"} (confidence interval coverage). Either
    (\code{estimate}, \code{se}) or (\code{lower}, \code{upper}) must be provided. \code{estimate} gives the name of the variable in \code{sim$results}
    containing the estimator of interest, \code{se} gives the name of the variable in \code{sim$results}
    containing the standard error of the estimator of interest, \code{lower} gives the name of the variable in
    \code{sim$results} containing the confidence interval lower bound, \code{upper} gives the name of the
    variable in \code{sim$results} containing the confidence interval upper bound, \code{truth} is the
    estimand of interest. See \emph{Details}.}

    \item{\code{stat = "is_na"}: \code{x} gives the name of the variable in \code{sim$results}
    on which to calculate the number of \code{NA}s.}
   }}
}
\value{
A data frame containing the result of each specified summary function as a column, for each of
    the simulation levels. The column \code{n_reps} returns the number of successful simulation replicates
    within each level.
}
\description{
This function calculates summary statistics for simulation results.
    Options for summary statistics include descriptive statistics (e.g. measures of
    center or spread) and inferential statistics (e.g. bias or confidence interval
    coverage). All summary statistics are calculated over simulation replicates
    within a single simulation level.
}
\details{
\itemize{

    \item{For all inferential summaries there are three ways to specify \code{truth}: (1) a single number,
    meaning the estimand is the same across all simulation replicates and levels, (2) a numeric vector of the
    same length as the number of rows in \code{sim$results}, or (3) the name of a variable in \code{sim$results}
    containing the estimand of interest.}

    \item{There are two ways to specify the confidence interval bounds for \code{coverage}. The first is to provide
    an \code{estimate} and its associated \code{se} (standard error). These should both be variables in
    \code{sim$results}. The function constructs a 95\% Wald-type confidence interval of the form
    (\code{estimate} - 1.96 \code{se}, \code{estimate} + 1.96 \code{se}).  The alternative is to provide
    \code{lower} and \code{upper} bounds, which should also be variables in \code{sim$results}. In this case,
    the confidence interval is (\code{lower}, \code{upper}). The coverage is simply the proportion of simulation
    replicates for a given level in which \code{truth} lies within the interval.}
}
}
\examples{
# The following is a toy example of a simulation, illustrating the use of
# the summarize function.
sim <- new_sim()
create_data <- function(n) { rpois(n, lambda=5) }
est_mean <- function(dat, type) {
  if (type=="M") { return(mean(dat)) }
  if (type=="V") { return(var(dat)) }
}
sim \%<>\% set_levels(n=c(10,100,1000), est=c("M","V"))
sim \%<>\% set_config(num_sim=5)
sim \%<>\% set_script(function() {
  dat <- create_data(L$n)
  lambda_hat <- est_mean(dat=dat, type=L$est)
  return (list("lambda_hat"=lambda_hat))
})
sim \%<>\% run()
sim \%>\% summarize(
  list(stat = "mean", name="mean_lambda_hat", x="lambda_hat"),
  list(stat = "mse", name="lambda_mse", estimate="lambda_hat", truth=5)
)
}
