% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.simba}
\alias{summary.simba}
\title{Summarize simulation results}
\usage{
\method{summary}{simba}(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{sd}{If \code{sd=TRUE} is passed, standard deviations are reported in
addition to means}

\item{coverage}{!!!!! TO DO}
}
\value{
!!!!! TO DO
}
\description{
Summarize simulation results
}
\examples{
!!!!! TO DO
}
