% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.simba_results}
\alias{summary.simba_results}
\title{Summarize simulation results}
\usage{
\method{summary}{simba_results}(sim_results, ...)
}
\arguments{
\item{sim_results}{A simulation results object created by run()}

\item{sd}{If `sd=TRUE` is passed, standard deviations are reported in
addition to means}
}
\value{
!!!!! TO DO
}
\description{
Summarize simulation results
}
\examples{
!!!!! TO DO
}
