% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update a simulation}
\usage{
update(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim(), that has already been run by run()}

\item{keep_errors}{Boolean (TRUE by default); if TRUE, do not try to re-run
simulation reps that results in errors previously; if FALSE, attempt to
run those reps again}

\item{keep_extra_reps}{Boolean (FALSE by default); if TRUE, keep previously run
reps even if they exceed the current num_sim in config; if FALSE, randomly
select excess reps to drop}

\item{keep_extra_levels}{Boolean (FALSE by default); if TRUE, keep previously run
levels even if they are no longer included in current sim levels; if FALSE,
drop them from results}
}
\description{
Update a simulation
}
\examples{
!!!!! TO DO
}
