% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_creator.R
\name{use_creator}
\alias{use_creator}
\title{Use a particular creator within a simulation}
\usage{
use_creator(creator, ...)
}
\value{
Runs the creator specified by the first argument within a simulation,
    supplying all additional arguments to the creator
}
\description{
Use a particular creator within a simulation
}
\examples{
!!!!! TO DO
}
