% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constants.R
\name{add_constants}
\alias{add_constants}
\title{Add simulation constants}
\usage{
add_constants(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object created by new_sim()}

\item{...}{One or more key-value pairs representing simulation constants
(any object that does not change across simulations).}
}
\value{
The original simulation object with the new constants added
}
\description{
Add simulation constants
}
\examples{
sim <- new_sim()
sim \%<>\% add_constants(
  "k" = 44,
  "tau" = c(0.25,0.5,0.75)
)
!!!!! continue example
}
