% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_creator.R
\name{add_creator}
\alias{add_creator}
\title{Add a "creator" function}
\usage{
add_creator(sim_obj, name, fn)
}
\arguments{
\item{sim_obj}{A simulation object created by new_sim()}

\item{name}{A name for the dataset-creating function}

\item{fn}{A function that creates a simulated dataset}
}
\value{
The original simulation object with the new creator function added
}
\description{
Add a "creator" function
}
\examples{
sim <- new_sim()
sim \%<>\% add_creator(
  "create_rct",
  function(n, sigma) {
    x <- runif(n)
    y <- 4*x + rnorm(n, mean=0, sd=sigma)
    return (data.frame(x=x,y=y))
  }
)
!!!!! continue example
}
