% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_method.R
\name{add_method}
\alias{add_method}
\title{Add a "method" function}
\usage{
add_method(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{name}{A name for the method function}

\item{fn}{A function that performs some analytic method. The function must
accept an object returned by a creator function as its first argument}
}
\value{
The original simulation object with the new method function added
}
\description{
Add a "method" function
}
\examples{
sim <- new_sim()
sim \%<>\% add_method(
  "OLS",
  function(data) {
    model <- lm(y~x, data=data)
    return (model$coefficients[["x"]])
  }
)
!!!!! continue example
}
