% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_script.R
\name{add_script}
\alias{add_script}
\title{Add a "simulation script" function}
\usage{
add_script(sim_obj, ...)
}
\value{
The original simulation object with the new script function added
}
\description{
Add a "simulation script" function
}
\examples{
!!!!! TO DO
!!!!! Script function should return a list of simple key-value pairs
!!!!! Need to modify if more complicated objects are returned
!!!!! Currently throws an error if there are NULL values in the list
!!!!! Make "return data frame" default option and "return list" as alternate
}
