% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{Wrapper to run a block of code as part of a batch}
\usage{
batch(code, warn = T)
}
\value{
TO DO
}
\description{
TO DO; note that you must create objects within batch(); do not change/delete existing objects
}
\examples{
TO DO
}
