% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sim.R
\name{config}
\alias{config}
\title{Modify simulation configuration}
\usage{
config(
  sim_obj,
  num_sim = NA,
  levels = NA,
  datasets = NA,
  parallel = NA,
  constants = NA
)
}
\arguments{
\item{sim_obj}{A simulation object returned by new_sim()}

\item{num_sim}{Number of simulations to conduct for each level combination}

\item{levels}{A list of vectors defining "simulation levels". If num_sim=5, 
then five simulations will be conducted for each combination of levels. 
For example, if levels=list("psi"=c(1,2), "eta"=c(8,9)), then five 
simulations will be conducted for (psi=1,eta=8), five for (psi=2,eta=8), 
five for (psi=1,eta=9), and five for (psi=2,eta=9).}

\item{datasets}{String; either "one" or "many". If set to "one", the same 
dataset will be used for all simulations. If set to "many", a new 
dataset will be generated for each simulation}

\item{parallel}{String; Either "inner", "outer", or "none".Controls which 
sections of the code are parallelized (i.e. for speed gains on 
multicore/cluster computing systems). Setting to "outer" will run one 
simulation per core. Setting to "inner" will parallelize operations 
within a single simulation. Setting to "none" will not parallelize any 
code (useful for debugging).}

\item{constants}{A list containing any objects (numbers, vectors, data 
frames) that remain constant across all simulations.}
}
\value{
A simulation object with modified configuration
}
\description{
Modify simulation configuration
}
\examples{
sim <- new_sim()
sim \%<>\% config(
  num_sim = 10,
  levels = list("rho"=c(0.1, 0.2), "eta"=c(2,3))
)
}
