% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complex.R
\name{get_complex}
\alias{get_complex}
\title{Access internal simulation variables}
\usage{
get_complex(sim, sim_uid)
}
\arguments{
\item{sim}{A simulation object of class \code{sim_obj}, usually created by
\code{\link{new_sim}}}

\item{sim_uid}{The unique identifier of a single simulation replicate. This
corresponds to the \code{sim_uid} column in \code{sim$results}.}
}
\value{
The value of the complex simulation result data corresponding to the
    supplied \code{sim_uid}, as well as the \code{rep_id}, \code{level_id},
    and associated level variables.
}
\description{
Extract complex simulation data from a simulation object
}
\examples{
sim <- new_sim()
sim \%<>\% set_levels(n=c(10, 100, 1000))
create_data <- function(n) {
  x <- runif(n)
  y <- 3 + 2*x + rnorm(n)
  return(data.frame("x"=x, "y"=y))
}
sim \%<>\% set_config(num_sim=2)
sim \%<>\% set_script(function() {
  dat <- create_data(L$n)
  model <- lm(y~x, data=dat)
  return(list(
    "beta0_hat" = model$coefficients[[1]],
    "beta1_hat" = model$coefficients[[2]],
    ".complex" = list(
      "model" = model,
      "cov_mtx" = vcov(model)
    )
  ))
})
sim \%<>\% run()
c5 <- get_complex(sim, sim_uid=5)
print(summary(c5$model))
print(c5$cov_mtx)
}
