% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sim.R
\name{new_sim}
\alias{new_sim}
\title{Initialize a new simulation object}
\usage{
new_sim()
}
\arguments{
\item{config}{A list. Contains simulation configuration information}
}
\value{
A simulation object
}
\description{
Initialize a new simulation object
}
\examples{
sim <- new_sim()
}
