% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R, R/run.R
\name{run}
\alias{run}
\title{Get the value of a simulation object variable}
\usage{
run(sim_obj, script, ...)

run(sim_obj, script, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{script}{The name of a simulation script, added using add_script()}

\item{variable}{The name of the internal variable}

\item{sim_uids}{A vector of sim_uids that represent simulations to run. If
omitted, all simulations are run. # update this !!!!!}
}
\description{
Get the value of a simulation object variable

Run the simulation
}
\examples{
!!!!! TO DO
!!!!! TO DO
}
