% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the simulation}
\usage{
run(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{sim_uids}{A vector of sim_uids that represent simulations to run. If
omitted, all simulations are run. # update this !!!!!}
}
\description{
Run the simulation
}
\examples{
!!!!! TO DO
}
