% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_on_cluster.R
\name{run_on_cluster}
\alias{run_on_cluster}
\title{Framework function for running simulations on a cluster computing system}
\usage{
run_on_cluster(first, main, last, config)
}
\arguments{
\item{first}{Code to run at the start of a simulation. This should be a
function that creates and returns a simulation object. Put everything you
need in the simulation object.}

\item{main}{Code that will run for every simulation replicate. This should
be a function that takes in your simulation object and returns the
results of a call to run()}

\item{last}{Code that will run after all simulation replicates have been run.
This should be a function that takes in your simulation object (which at
this point will contain your results) and does something with it (e.g.
displays the results on a graph).}

\item{config}{A list of configuration options. !!!!! TO DO}
}
\description{
Framework function for running simulations on a cluster computing system
}
\examples{
!!!!! TO DO
}
