% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Modify the simulation configuration}
\usage{
set_config(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class \code{simba}, usually created by
\link{new_sim}}

\item{num_sim}{Number of simulations to conduct for each level combination}

\item{datasets}{String; either "one" or "many". If set to "one", the same
dataset will be used for all simulations. If set to "many", a new
dataset will be generated for each simulation}

\item{parallel}{String; Either "inner", "outer", or "none".Controls which
sections of the code are parallelized (i.e. for speed gains on
multicore/cluster computing systems). Setting to "outer" will run one
simulation per core. Setting to "inner" will parallelize operations
within a single simulation. Setting to "none" will not parallelize any
code (useful for debugging).}

\item{packages}{A character vector of packages to load}

\item{stop_at_error}{Boolean (FALSE by default); if TRUE, simulation will
stop after it encounters an error in any single run. Useful for
debugging.}

\item{dir}{Directory (given as a character string) where simulation files
should be stored; defaults to current working directory}
}
\value{
The original simulation object with a modified configuration
}
\description{
This function sets configuration options for the simulation. If
    the 'packages' argument is specified, all packages will be loaded and
    attached via library() when set_config() is called. Multiple calls to
    set_config() will only overwrite configuration options that are specified
    in the subsequent calls, leaving others in place.
}
\examples{
sim <- new_sim()
sim \%<>\% set_config(
  num_sim = 10
)
}
