% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_levels.R
\name{set_levels}
\alias{set_levels}
\title{Set simulation levels}
\usage{
set_levels(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{...}{One or more key-value pairs representing simulation levels (see
below)}
}
\value{
The original simulation object with the old set of levels replaced
    with the new set
}
\description{
Set simulation levels
}
\examples{
sim <- new_sim()
sim \%<>\% set_levels(
  "rho" = c(0.1, 0.2),
  "eta" = c(2,3)
)
!!!!! continue example
}
