% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_script.R
\name{set_script}
\alias{set_script}
\title{Add a "simulation script" function}
\usage{
set_script(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim()}

\item{fn}{A function !!!!!}
}
\value{
The original simulation object with the new script function added.
The script should be a function that returns a list of key-value pairs. !!!!! continue
}
\description{
Add a "simulation script" function
}
\examples{
!!!!! TO DO
}
