% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimEngine-package.R
\docType{package}
\name{SimEngine-package}
\alias{SimEngine}
\alias{SimEngine-package}
\title{SimEngine: A Modular Framework for Statistical Simulations in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An open-source R package for structuring, maintaining, running, and debugging statistical simulations on both local and cluster-based computing environments. Emphasis is placed on thorough documentation and scalability. See full documentation at \url{https://avi-kenny.github.io/SimEngine/}.
}
\author{
\strong{Maintainer}: Avi Kenny \email{avi.kenny@gmail.com}

Authors:
\itemize{
  \item Charles Wolock \email{cwolock@uw.edu}
}

}
\keyword{internal}
