% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update a simulation}
\usage{
update(sim_obj, keep_errors = TRUE, keep_extra = FALSE)
}
\arguments{
\item{sim_obj}{A simulation object of class "simba", usually created by
new_sim(), that has already been run by run()}

\item{keep_errors}{Boolean (TRUE by default); if TRUE, do not try to re-run
simulation reps that results in errors previously; if FALSE, attempt to
run those reps again}

\item{keep_extra}{Boolean (FALSE by default); if TRUE, keep previously run
simulation reps even if they exceed the current num_sim in config or are from
a level that has been dropped; if FALSE, drop excess reps (starting from the last rep
for that particular simulation level)}
}
\description{
Update a simulation
}
\examples{
!!!!! TO DO
}
