% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_method.R
\name{use_method}
\alias{use_method}
\title{Use a particular method within a simulation}
\usage{
use_method(method, ...)
}
\value{
Runs the method specified by the first argument within a simulation,
    supplying all additional arguments to the method
}
\description{
Use a particular method within a simulation
}
\examples{
!!!!! TO DO
}
